# Description:
Added an example for accessing entities from a MongoDB - still work in
progress...
Also improved handling of URLs. 

## Type of change:
<!-- Put an `x` in the box that applies. -->
- [x] Bug fix.
- [x] New feature.
- [ ] Documentation update.

## Checklist for the reviewer:
<!-- Put an `x` in the boxes that apply. These can be filled by reviewer
after the PR is created. -->

This checklist should be used as a help for the reviewer.

- [ ] Is the change limited to one issue?
- [ ] Does this PR close the issue?
- [ ] Is the code easy to read and understand?
- [ ] Do all new feature have an accompanying new test?
- [ ] Has the documentation been updated as necessary?

+++ storages/python/dlite-plugins-python.c
@@ -366,7 +366,7 @@ int memsaver(const DLiteStoragePlugin *api, unsigned char *buf, size_t size,
   }
   assert(length > 0);
   memcpy(buf, buffer, (size > (size_t)length) ? (size_t)length : size);
-  retval = length;
+  retval = (int)length;
  fail:
   Py_XDECREF(pyinst);
   Py_XDECREF(v);
